python-playwright/requirements.txt
pytest-playwright == 0.5.2
fastapi == 0.114.1
